+++ src/image/ImageBrightness.java
@@ -10,9 +10,9 @@ import java.awt.*;
  * @author Or Tarazi, Agam Hershko
  */
 public class ImageBrightness {
-    private static final float RED_FACTOR = 0.2126f;
-    private static final float GREEN_FACTOR = 0.7152f;
-    private static final float BLUE_FACTOR = 0.0722f;
+    private static final double RED_FACTOR = 0.2126;
+    private static final double GREEN_FACTOR = 0.7152;
+    private static final double BLUE_FACTOR = 0.0722;
     private static final int MAX_RGB_VALUE = 255;
 
     /**
@@ -20,10 +20,10 @@ public class ImageBrightness {
      *              values and returns the average, which is the equivalent brightness value of the sub-image.
      * @return the calculated brightness value of a sub-image
      */
-    public static float calculateImageBrightness(Image image) {
+    public static double calculateImageBrightness(Image image) {
         int width = image.getWidth();
         int height = image.getHeight();
-        float grayScaleTotal = 0;
+        double grayScaleTotal = 0;
 
         for (int row = 0; row < height; row++) {
             for (int col = 0; col < width; col++) {
@@ -41,7 +41,7 @@ public class ImageBrightness {
      * @param color an RGB color.
      * @return grayscale value.
      */
-    private static float colorToGrayscale(Color color) {
+    private static double colorToGrayscale(Color color) {
         int red = color.getRed();
         int green = color.getGreen();
         int blue = color.getBlue();
