+++ src/image_char_matching/CharBrightnessMap.java
@@ -16,15 +16,15 @@ import ascii_art.RoundType;
  * @author Or Tarazi, Agam Hershko
  */
 public class CharBrightnessMap {
-    private static final float MIN_POSSIBLE_BRIGHTNESS = 0;
-    private static final float MAX_POSSIBLE_BRIGHTNESS = 1;
+    private static final double MIN_POSSIBLE_BRIGHTNESS = 0;
+    private static final double MAX_POSSIBLE_BRIGHTNESS = 1;
 
     // Stores raw brightness values
-    private final TreeMap<Integer, Float> rawBrightnessMap;
+    private final TreeMap<Integer, Double> rawBrightnessMap;
     // Stores normalized brightness values
-    private final TreeMap<Integer, Float> normalizedBrightnessMap;
-    private float minBrightness;
-    private float maxBrightness;
+    private final TreeMap<Integer, Double> normalizedBrightnessMap;
+    private double minBrightness;
+    private double maxBrightness;
 
     /**
      * constructor for the data-structure. initializes two TreeMap when an instance is created.
@@ -43,7 +43,7 @@ public class CharBrightnessMap {
      * @param c char to be added to ascii set
      */
     public void addChar(char c) {
-        float brightness = CharBrightnessCalculator.calculateCharBrightness(c);
+        double brightness = CharBrightnessCalculator.calculateCharBrightness(c);
         this.rawBrightnessMap.put((int) c, brightness); // Always store raw brightness
 
         if (brightness > this.maxBrightness || brightness < this.minBrightness) {
@@ -70,17 +70,17 @@ public class CharBrightnessMap {
      *                   - DOWN: Select the closest character with brightness below or equal to given value.
      * @return the ASCII character that best matches the brightness value based on the rounding type.
      */
-    public char getCharByNormalizedBrightness(float brightness, RoundType roundType) {
-        Map.Entry<Integer, Float> closestAbove = null;
-        Map.Entry<Integer, Float> closestBelow = null;
+    public char getCharByNormalizedBrightness(double brightness, RoundType roundType) {
+        Map.Entry<Integer, Double> closestAbove = null;
+        Map.Entry<Integer, Double> closestBelow = null;
 
-        // Initialize with Float.MAX_VALUE to ensure any valid difference is smaller
-        float smallestDifferenceAbove = Float.MAX_VALUE;
-        float smallestDifferenceBelow = Float.MAX_VALUE;
+        // Initialize with double.MAX_VALUE to ensure any valid difference is smaller
+        double smallestDifferenceAbove = Double.MAX_VALUE;
+        double smallestDifferenceBelow = Double.MAX_VALUE;
 
         // Single loop to find both the closestAbove and closestBelow entries
-        for (Map.Entry<Integer, Float> entry : this.normalizedBrightnessMap.entrySet()) {
-            float difference = entry.getValue() - brightness;
+        for (Map.Entry<Integer, Double> entry : this.normalizedBrightnessMap.entrySet()) {
+            double difference = entry.getValue() - brightness;
             if (difference == 0) { // Exact match: Return immediately
                 return (char) (int) entry.getKey();
             } else if (difference > 0 && difference < smallestDifferenceAbove) { // Check for closest above
@@ -111,9 +111,9 @@ public class CharBrightnessMap {
      *                     - DOWN: Select the closest character with brightness below or equal to brightness.
      * @return the ASCII character that matches the target brightness based on the rounding strategy.
      */
-    private static char getRoundedChar(Map.Entry<Integer, Float> closestAbove,
-                                       Map.Entry<Integer, Float> closestBelow,
-                                       float brightness, RoundType roundType) {
+    private static char getRoundedChar(Map.Entry<Integer, Double> closestAbove,
+                                       Map.Entry<Integer, Double> closestBelow,
+                                       double brightness, RoundType roundType) {
         return switch (roundType) {
             case ABS -> {
                 if (closestAbove != null && closestBelow != null) {
@@ -142,7 +142,7 @@ public class CharBrightnessMap {
      * @param c key in map to be removed
      */
     public void removeChar(char c) {
-        float brightness = this.rawBrightnessMap.remove((int) c);
+        double brightness = this.rawBrightnessMap.remove((int) c);
         this.normalizedBrightnessMap.remove((int) c);
 
         if (brightness == this.maxBrightness || brightness == this.minBrightness) {
@@ -162,7 +162,7 @@ public class CharBrightnessMap {
         this.minBrightness = MAX_POSSIBLE_BRIGHTNESS;
         this.maxBrightness = MIN_POSSIBLE_BRIGHTNESS;
 
-        for (float brightness : this.rawBrightnessMap.values()) {
+        for (double brightness : this.rawBrightnessMap.values()) {
             this.adjustMinAndMaxBrightness(brightness);
         }
     }
@@ -172,8 +172,8 @@ public class CharBrightnessMap {
      * it's not needed to be called when minimum/brightness haven't changed after adding an ascii char.
      */
     private void normalizeAllValues() {
-        for (Map.Entry<Integer, Float> entry : this.rawBrightnessMap.entrySet()) {
-            float normalized = this.normalizeBrightness(entry.getValue());
+        for (Map.Entry<Integer, Double> entry : this.rawBrightnessMap.entrySet()) {
+            double normalized = this.normalizeBrightness(entry.getValue());
             this.normalizedBrightnessMap.put(entry.getKey(), normalized);
         }
     }
@@ -187,7 +187,7 @@ public class CharBrightnessMap {
      * @param addedBrightness the new brightness value in the map that changes the min/max values.
      * @return normalized brightness
      */
-    private float normalizeBrightness(float addedBrightness) {
+    private double normalizeBrightness(double addedBrightness) {
         return (addedBrightness - this.minBrightness) / (this.maxBrightness - this.minBrightness);
     }
 
@@ -198,7 +198,7 @@ public class CharBrightnessMap {
      *
      * @param addedBrightness the new brightness value in the map that changes the min/max values.
      */
-    private void adjustMinAndMaxBrightness(float addedBrightness) {
+    private void adjustMinAndMaxBrightness(double addedBrightness) {
         if (addedBrightness < this.minBrightness) {
             this.minBrightness = addedBrightness;
         }
